Samples refactor and new docs for samples
