+++ src/main/java/org/mybatis/spring/mapper/MapperFactoryBean.java
@@ -59,7 +59,7 @@ public class MapperFactoryBean<T> extends SqlSessionDaoSupport implements Factor
   private boolean addToConfig = true;
 
   public MapperFactoryBean() {
-    //intentially empty 
+    //intentionally empty 
   }
   
   public MapperFactoryBean(Class<T> mapperInterface) {
