+++ src/test/java/org/mybatis/spring/sample/config/applicationContext-infrastructure.xml
@@ -56,11 +56,11 @@
              file. This config file could, in turn, contain &ltmapper&gt
              elements that point to the correct mapper xml files.
          -->
-        <property name="mapperLocations" value="classpath:org/mybatis/spring/sample/dao/*.xml" />
+        <property name="mapperLocations" value="classpath:org/mybatis/spring/sample/mapper/*.xml" />
     </bean>
     
-    <!-- simple transactional service layer bean; the userDao can be implemented in various ways with MyBatis-Spring  -->
+    <!-- simple transactional service layer bean; MyBatis-Spring will create a proxy for the userMapper bean -->
     <bean id="fooService" class="org.mybatis.spring.sample.service.FooService">
-        <property name="userDao" ref="userDao" />
+        <property name="userMapper" ref="userMapper" />
     </bean>
 </beans>
